--- InstProfile/Posts&Stories/Constants/StoryConstants.swift
@@ -1,18 +1,0 @@
-//
-//  Constants.swift
-//  InstProfile
-//
-//  Created by Владимир Есаян on 21.09.2023.
-//
-
-import UIKit
-
-// MARK: - add photos fot Story
-struct storyImages {
-    static let ufo        = UIImage(named: "13")!
-    static let starship   = UIImage(named: "14")!
-    static let mars       = UIImage(named: "15")!
-    static let spaceRover = UIImage(named: "16")!
-    static let telescope  = UIImage(named: "17")!
-    static let starlink   = UIImage(named: "18")!
-}
