+++ InstProfile/Stories& Posts/Constants/StoryConstants.swift
@@ -1,0 +1,18 @@
+//
+//  Constants.swift
+//  InstProfile
+//
+//  Created by Владимир Есаян on 21.09.2023.
+//
+
+import UIKit
+
+// MARK: - add photos fot Story
+struct storyImages {
+    static let story1 = UIImage(named: "13")!
+    static let story2 = UIImage(named: "14")!
+    static let story3 = UIImage(named: "15")!
+    static let story4 = UIImage(named: "16")!
+    static let story5 = UIImage(named: "17")!
+    static let story6 = UIImage(named: "18")!
+}
